requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
ddt==1.4.4
et-xmlfile==1.1.0
Faker==11.3.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
jsonpath==0.82
numpy==1.22.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.6
PySocks==1.7.1
pytest==7.1.2
pytest-assume==2.4.3
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-sugar==0.9.4
pytest_check==1.0.4
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.1.0
ddddocr==1.4.7
cx_Oracle==8.3.0